# assets/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Athena HTTP JSON APIs. API clients can authenticate protected operations with an Athena API Key by sending Authorization: Bearer <token>. The Athena server separately evaluates current account entitlements and any required module authorization for each protected request; this Swagger security declaration describes credential transport only and does not grant access.",
    "title": "Athena API",
    "version": "v1.0.0-1067-gfa7b246e-dirty"
  },
  "paths": {
...
    "/api/v1/wallets": {
      "get": {
        "tags": [
          "WalletService"
        ],
        "operationId": "WalletService_ListWallets",
        "parameters": [
          {
            "type": "string",
            "name": "walletType",
            "in": "query"
          },
          {
            "type": "string",
            "name": "query",
            "in": "query"
          },
          {
            "type": "integer",
            "format": "int32",
            "name": "page",
            "in": "query"
          },
          {
            "type": "integer",
            "format": "int32",
            "name": "pageSize",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/walletListWalletsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        }
      },
      "post": {
        "tags": [
          "WalletService"
        ],
        "operationId": "WalletService_CreateWallet",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/walletCreateWalletRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/walletCreateWalletResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        }
      }
    },
    "/api/v1/wallets/{id}": {
      "get": {
        "tags": [
          "WalletService"
        ],
        "operationId": "WalletService_GetWallet",
        "parameters": [
          {
            "type": "string",
            "format": "int64",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/walletGetWalletResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        }
      }
    },
...
    "/api/v1/wallets:import": {
      "post": {
        "tags": [
          "WalletService"
        ],
        "operationId": "WalletService_ImportWallet",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/walletImportWalletRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/walletImportWalletResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "type_url": {
          "type": "string"
        },
        "value": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "runtimeError": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        },
        "error": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
...
    "v1alpha1WalletItem": {
      "type": "object",
      "properties": {
        "address": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "remark": {
          "type": "string"
        },
        "revision": {
          "type": "integer",
          "format": "uint64"
        },
        "source": {
          "type": "string"
        },
        "walletType": {
          "type": "string"
        },
        "avatarKind": {
          "type": "string"
        },
        "avatarPresetId": {
          "type": "string"
        },
        "avatarUrl": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        }
      }
    },
...
    "walletCreateWalletRequest": {
      "type": "object",
      "properties": {
        "remark": {
          "type": "string"
        },
        "walletType": {
          "type": "string"
        },
        "avatarPresetId": {
          "type": "string"
        }
      }
    },
    "walletCreateWalletResponse": {
      "type": "object",
      "properties": {
        "item": {
          "$ref": "#/definitions/v1alpha1WalletItem"
        },
        "privateKey": {
          "type": "string"
        }
      }
    },
    "walletGetWalletResponse": {
      "type": "object",
      "properties": {
        "item": {
          "$ref": "#/definitions/v1alpha1WalletItem"
        }
      }
    },
    "walletImportWalletRequest": {
      "type": "object",
      "properties": {
        "remark": {
          "type": "string"
        },
        "walletType": {
          "type": "string"
        },
        "privateKey": {
          "type": "string"
        },
        "avatarPresetId": {
          "type": "string"
        }
      }
    },
    "walletImportWalletResponse": {
      "type": "object",
      "properties": {
        "item": {
          "$ref": "#/definitions/v1alpha1WalletItem"
        }
      }
    },
    "walletListWalletsResponse": {
      "type": "object",
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1alpha1WalletItem"
          }
        },
        "page": {
          "type": "integer",
          "format": "int32"
        },
        "total": {
          "type": "integer",
          "format": "int64"
        },
        "pageSize": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "athenaBearer": {
      "description": "An Athena API Key supplied as Authorization: Bearer <token>. The account must have API Key access enabled and satisfy the server-side permission required by the operation.",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "athenaBearer": []
    }
  ]
}